Revert "Move username/password tuples from server state to an ETS table"
This reverts commit 37f44d03a08a54d68874ce9880c2dfbcf946f81b.

There are many flaws with this patch. First, the ETS table used to store
user/password tuples is always created during the configuration parsing (or
during the loading in embedded mode). So during a configuration reload, the old
server configurations and the new one never match because the table id in
used, when Yaws configuration is reloaded, the servers are restarted, even when
there is no changes.

Then, the ETS table is owned by the process that parses the configuration (or
that loads it in embedded mode). So the table is deleted when this process
dies. This happens when the configuration is reloaded using the startup
script. This is a huge bug, because after a reload, all requests requiring a
user/password authentication will crash because the table is gone.

And finally, ETS table owned by old server configurations is never garbaged.