Rewrite yaws_compile module to improve .yaws script parsing
yaws_compile modules has been rewritten from scratch to be more flexible. Now,
it can handle following syntaxes:

    <erl>out(_A) -> {html, "test1"}. </erl>

    <b><erl>out(_A) -> {html, "test2"}. </erl></b>

    <b><erl>
      out(_A) -> {html, "test3"}.
    </erl></b>

In addition, it support strange tag syntaxes with spaces and newlines, like:

    <erl
       module =
         "my_module"
     >
       out(_A) -> {html, "test"}.
    </erl
          >

The sames syntaxes are supported for verbatim blocks of course.

Unknown attributes in erl tags (all except "module") are ignored instead of
crashing the Yaws process. When a module name is provided for an erl block, we
check it to be sure that it will not override an existing module (owned by
another yaws script or a "real" module, like yaws !).

About verbatim blocks, it is now possible to add HTML attributes that will be
keep after substitution. For example:

     <verbatim class="code" id="listing-1">
        ...
     </verbatim>

will be replaced by:

     <pre class="code" id="listing-1">
        ...
     </pre>

Finally, about performance, the new parser tends to be faster in many cases when
one file is compiled. But it is really faster when many scripts are compiled in
parallel. The old parser used a global lock on the node. So when a script was
parsed, all others was blocked. Now, a lock is set per script. So two different
script can be parsed in same time.